Move freetype image in a separate directive
